ParentApp/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
oauth2client==4.1.3
Pillow==9.0.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive2==1.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyYAML==6.0
requests==2.27.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
